docs/theme/sphinx-material/requirements.txt
sphinx>=2.0
beautifulsoup4
python-slugify[unidecode]
css_html_js_minify
lxml
